Provide access to the TID of the runStmt sandbox
We were emulating this inside ide-backend itself, but that meant that we might
send a UserInterrupt to ghc while compiling the statement rather than to the
user code, with all kinds of horrid results (exception ending up as part of
some thunk due to some unsafePerformIO, thrown every time we touched that
thunk?). Hard to reproduce and even harder to figure out what's going on
precisely. My moving this to inside ghc we hopefully avoid this problem.