Mark the source with RELEASE=NO to get more informative ghc version
It does not help to tell 2 versions from the same day, but so far
we are not churning GHC patches so often. Note that you need
to wipe the codebase with 'make clean' and 'perl boot' for this
to take effect. Also, this may interfere with your sandboxes, e.g.,
a ~/env/active/local/bin/ghc-7.4.2.20130508 link may need to be added.